+++ packages/react-magma-dom/src/components/Select/SelectTriggerButton.tsx
@@ -48,6 +48,7 @@ export function SelectTriggerButton<T>(props: SelectTriggerButtonInterface<T>) {
     customComponents,
     hasError,
     disabled,
+    isInverse,
     style: passedInStyle,
     toggleButtonProps,
   } = props;
@@ -66,7 +67,7 @@ export function SelectTriggerButton<T>(props: SelectTriggerButtonInterface<T>) {
       data-testid="selectTriggerButton"
       hasError={hasError}
       disabled={disabled}
-      isInverse={false}
+      isInverse={isInverse}
       role="button"
       style={style}
       theme={theme}