+++ packages/react-magma-dom/src/index.ts
@@ -119,6 +119,7 @@ export {
 } from './components/LoadingIndicator';
 export { NavTabs } from './components/NavTabs';
 export { NavTab } from './components/NavTabs/NavTab';
+export { Pagination } from './components/Pagination';
 export { Paragraph, ParagraphProps } from './components/Paragraph';
 export { PasswordInput, PasswordInputProps } from './components/PasswordInput';
 export {
@@ -213,4 +214,8 @@ export { I18nContext } from './i18n';
 export { defaultI18n } from './i18n/default';
 export { I18nInterface } from './i18n/interface';
 export { useMediaQuery } from './hooks/useMediaQuery';
+export { useDataPagination } from './hooks/useDataPagination';
 export { InverseContext, useIsInverse } from './inverse';
+
+export * from './components/Datagrid';
+export * from './components/Transition';