+++ website/react-magma-docs/src/components/DarkMode/DarkModeContext.js
@@ -1,0 +1,6 @@
+import * as React from 'react';
+
+export const DarkModeContext = React.createContext({
+  isDarkMode: false,
+  setIsDarkMode: () => {},
+});